Don't log stdout when running 'macrobe list'
The 'macrobe list' command's output is useful in scripts that won't
necessarily be attached to a tty.

For example, when running 'macrobe list | rofi -dmenu' from a rofi
menu (or a blezz menu, or an i3/sxhkd keybinding, etc).

Also, add an example for launching a rofi menu to the example
blezz config.